backend/requirements.txt
fastapi
uvicorn
sqlalchemy
paho-mqtt
google-generativeai
python-dotenv
psycopg2-binary
